switch bug test
